user_authentication/requirements.txt
azure-functions
pymysql==5.0.1